requirements.txt
Django==4.1.4
django-cleanup==6.0.0
djangorestframework==3.14.0
django-cloudinary-storage
dj-static